refactor: extract utility functions from component file to utils directory
