+++ components/slider/PropsType.tsx
@@ -1,4 +1,4 @@
-import { ReactNode } from 'react';
+import type { ReactNode } from 'react';
 
 export interface SliderMarks {
   [key: number]: ReactNode;
