+++ components/slider/utils/getPrecision.ts
@@ -1,0 +1,8 @@
+export default function getPrecision(step: number) {
+  const stepString = step.toString();
+  let precision = 0;
+  if (stepString.indexOf('.') >= 0) {
+    precision = stepString.length - stepString.indexOf('.') - 1;
+  }
+  return precision;
+}
