+++ components/slider/utils/preventDefault.ts
@@ -1,0 +1,3 @@
+export default function preventDefault(event: MouseEvent) {
+  event.preventDefault();
+}
